Fix refactoring 2
